requirements.txt
streamlit
spacy
st-annotated-text